Pagina Final

+++ src/app/services/lenguajes.service.ts
@@ -1,6 +1,9 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, throwError } from 'rxjs';
+import { catchError } from 'rxjs/operators';
+import { AuthService } from './auth.service';
+
 
 @Injectable({
   providedIn: 'root'
@@ -8,9 +11,8 @@ import { Observable } from 'rxjs';
 export class LenguajesService {
 
   url: string ="https://apirest-2fb0f-default-rtdb.firebaseio.com/v17topList";
-  apiurl: string = "https://paises-a8a38-default-rtdb.firebaseio.com/"
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient, private authService: AuthService) { }
 
   getListlanguages(): Observable<any> 
   {
@@ -20,19 +22,50 @@ export class LenguajesService {
 
   postlanguages(body:any): Observable<any>
   {
-    let postUrl = this.url + ".json"
-    return this.http.post(postUrl, body) 
+       if (this.authService.getuser() !== null) {
+        let postUrl = this.url + ".json"
+        return this.http.post(postUrl, body) 
+        .pipe(
+          catchError(error => this.handleError(error))
+        );
+    } else {
+      // Manejar el caso cuando el usuario no está autenticado
+      return throwError('Usuario no autenticado');
+    }
+    
+   
   }
 
   deletelanguages(id:string): Observable<any>
   {
-    let delUrl = this.url + "/" + id + ".json"
-    return this.http.delete(delUrl)
+    if (this.authService.getuser() !== null) {
+      let delUrl = this.url + "/" + id + ".json"
+      return this.http.delete(delUrl)
+        .pipe(
+          catchError(error => this.handleError(error))
+        );
+    } else {
+      // Manejar el caso cuando el usuario no está autenticado
+      return throwError('Usuario no autenticado');
+    }
+    
   }
 
   updateLanguage(id:string, body:any): Observable<any>
   {
-    let uptUrl = this.url + "/" + id + ".json"
-    return this.http.put(uptUrl, body) 
+    if (this.authService.getuser() !== null) {
+      let uptUrl = this.url + "/" + id + ".json"
+      return this.http.put(uptUrl, body)
+        .pipe(
+          catchError(error => this.handleError(error))
+        );
+    } else {
+      // Manejar el caso cuando el usuario no está autenticado
+      return throwError('Usuario no autenticado');
+    }
+  }
+  private handleError(error: any): Observable<never> {
+    console.error('Error in LenguajesService:', error);
+    return throwError('Something went wrong, please try again later.');
   }
 }
